LiveviewRecipesWeb.CoreComponents (LiveView Recipes v0.1.0)

Provides core UI components.

The components in this module use Tailwind CSS, a utility-first CSS framework. See the Tailwind CSS documentation to learn how to customize the generated components in this module.

Icons are provided by heroicons, using the heroicons_elixir project.

Functions

back(assigns)

Renders a back navigation link.

Examples

<.back navigate={~p"/posts"}>Back to posts</.back>

Attributes

navigate (:any) (required)

Slots

inner_block (required)

button(assigns)

Renders a button.

Examples

<.button>Send!</.button>
<.button phx-click="go" class="ml-2">Send!</.button>

Attributes

type (:string) - Defaults to nil.

class (:string) - Defaults to nil. Global attributes are accepted.

Slots

inner_block (required)

error(assigns)

Generates a generic error message.

Slots

inner_block (required)

flash(assigns)

Renders flash notices.

Examples

<.flash kind={:info} flash={@flash} />
<.flash kind={:info} phx-mounted={show("#flash")}>Welcome Back!</.flash>

Attributes

id (:string) - the optional id of flash container. Defaults to "flash".

flash (:map) - the map of flash messages to display. Defaults to %{}.

title (:string) - Defaults to nil.

kind (:atom) - used for styling and flash lookup.Must be one of :info, or :error.

autoshow (:boolean) - whether to auto show the flash on mount. Defaults to true.

close (:boolean) - whether the flash can be closed. Defaults to true. Global attributes are accepted.

Slots

inner_block - the optional inner block that renders the flash message.

header(assigns)

Renders a header with title.

Attributes

class (:string) - Defaults to nil.

Slots

inner_block (required)

subtitle

actions

hide(js \\ %JS{}, selector)

hide_modal(js \\ %JS{}, id)

input(assigns)

Renders an input with label and error messages.

A %Phoenix.HTML.Form{} and field name may be passed to the input to build input names and error messages, or all the attributes and errors may be passed explicitly.

Examples

<.input field={{f, :email}} type="email" />
<.input name="my-input" errors={["oh no!"]} />

Attributes

id (:any)

name (:any)

label (:string) - Defaults to nil.

type (:string) - Defaults to "text".

value (:any)

field (:any) - a %Phoenix.HTML.Form{}/field name tuple, for example: {f, :email}.

errors (:list)

checked (:boolean) - the checked flag for checkbox inputs.

prompt (:string) - the prompt for select inputs. Defaults to nil.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2.

multiple (:boolean) - the multiple flag for select inputs. Defaults to false. Global attributes are accepted.

Slots

inner_block

label(assigns)

Renders a label.

Attributes

for (:string) - Defaults to nil.

Slots

inner_block (required)

list(assigns)

Renders a data list.

Examples

<.list>
  <:item title="Title"><%= @post.title %></:item>
  <:item title="Views"><%= @post.views %></:item>
</.list>

Slots

item (required) - Accepts attributes:

title (:string) (required)

modal(assigns)

Renders a modal.

Examples

<.modal id="confirm-modal">
  Are you sure?
  <:confirm>OK</:confirm>
  <:cancel>Cancel</:cancel>
</.modal>

JS commands may be passed to the :on_cancel and on_confirm attributes for the caller to react to each button press, for example:

<.modal id="confirm" on_confirm={JS.push("delete")} on_cancel={JS.navigate(~p"/posts")}>
  Are you sure you?
  <:confirm>OK</:confirm>
  <:cancel>Cancel</:cancel>
</.modal>

Attributes

id (:string) (required)

show (:boolean) - Defaults to false.

on_cancel (Phoenix.LiveView.JS) - Defaults to %Phoenix.LiveView.JS{ops: []}.

on_confirm (Phoenix.LiveView.JS) - Defaults to %Phoenix.LiveView.JS{ops: []}.

Slots

inner_block (required)

title

subtitle

confirm

cancel

show(js \\ %JS{}, selector)

show_modal(js \\ %JS{}, id)

simple_form(assigns)

Renders a simple form.

Examples

<.simple_form :let={f} for={:user} phx-change="validate" phx-submit="save">
  <.input field={{f, :email}} label="Email"/>
  <.input field={{f, :username}} label="Username" />
  <:actions>
    <.button>Save</.button>
  </:actions>
</.simple_form>

Attributes

for (:any) - the datastructure for the form. Defaults to nil.

as (:any) - the server side parameter to collect all input under. Defaults to nil. Global attributes are accepted.

Slots

inner_block (required)

actions - the slot for form actions, such as a submit button.

table(assigns)

Renders a table with generic styling.

Examples

<.table id="users" rows={@users}>
  <:col :let={user} label="id"><%= user.id %></:col>
  <:col :let={user} label="username"><%= user.username %></:col>
</.table>

Attributes

id (:string) (required)

row_click (:any) - Defaults to nil.

rows (:list) (required)

Slots

col (required) - Accepts attributes:

label (:string)

action - the slot for showing user actions in the last table column.

translate_error(arg)

Translates an error message using gettext.

translate_errors(errors, field)

Translates the errors for a field from a keyword list of errors.